Тесты | вывод ошибки, при попытки добавления эпизода, если нет сезонов

+++ app/Http/Controllers/FavouriteController.php
@@ -17,16 +17,20 @@ class FavouriteController extends Controller
     }
 
     public function add_favourites($anime_id): RedirectResponse {
-        $favourite = Favourite::where('user_id', Auth::user()->id)->where('anime_id', $anime_id)->first();
-        if (is_null($favourite)) {
-            Favourite::create([
-                'user_id' => Auth::user()->id,
-                'anime_id' => $anime_id,
-            ]);
+        if (Anime::where('id', $anime_id)->firstOrFail()){
+            $favourite = Favourite::where('user_id', Auth::user()->id)->where('anime_id', $anime_id)->first();
+            if (is_null($favourite)) {
+                Favourite::create([
+                    'user_id' => Auth::user()->id,
+                    'anime_id' => $anime_id,
+                ]);
+            } else {
+                Favourite::find($favourite->id)->delete();
+            }
+            return redirect()->back();
         } else {
-            Favourite::find($favourite->id)->delete();
+            abort('404');
         }
-        return redirect()->back();
     }
 
     public function destroy(Request $request): RedirectResponse {
